Started to add the new error handling.
FIXME: This is not done yet! Atm, it seems to works, but every error-message from the target function is surpressed! If no imputation function is given, every error generates a "Infeasible y" error. Wen want to save the real error message in the optimization path, this is not done yet.